@@ -17,7 +17,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <javaVersion>1.7</javaVersion>
+        <javaVersion>1.8</javaVersion>
         <!--Possibility to deploy directly to the plugins folder-->
         <outputDir>${basedir}/target</outputDir>
     </properties>
